Make the changelog conform to the keepachangelog.com style (#172)
The changelog intro says it conforms to the https://keepachangelog.com style, but previously the content actually did not. Now it does match that style, making it consistent with our other CNBs.

See:
https://keepachangelog.com/en/1.1.0/

I've also added links to the GitHub PRs for each entry. Whilst adding these is extra work for each PR, I feel it's worthwhile for end-users -- I very much find it useful when projects we use do the same.

GUS-W-14283439.