Add buildpack description to buildpack.toml
Adds the buildpack `description` to `buildpack.toml` with
a style matching that discussed in:
https://github.com/heroku/cnb-builder-images/issues/408

The `description` field is displayed on the CNB registry's page for
the buildpack.

In addition, I've cleaned up the `[metadata]` table/subtables to match
the more concise style used by the Procfile and Python CNBs.

GUS-W-14121598.