Switch to Cargo.toml based lint configuration
As of the Cargo included in Rust 1.74, lints can now be configured in
`Cargo.toml` across whole crates/workspaces:
https://blog.rust-lang.org/2023/11/16/Rust-1.74.0.html
https://doc.rust-lang.org/stable/cargo/reference/manifest.html#the-lints-section

This reduces the boilerplate, and the chance that we forget to enable
lints in some targets.

Since this feature requires Rust 1.74, the MSRV has also been bumped
(however, libcnb 0.16.0 already requires Rust 1.74, so in practice
this is a no-op).

GUS-W-14523792.