Merge Artifact version and semantic_version fields
This comes at the cost of losing information about the original string from the source. However, the actual value is still contained in the url field (and can be included depending on the implementation of "Version" for Artifact<V> caller, and seems like a reasonable trade-off considering the information is still contained in the artifact.