Fix issues with Docker Hub rate limits in CI
Similar to the Node.js CNB repo, this repo is occasionally hitting
Docker Hub rate limit errors on the ARM jobs in the matrix like so:
https://github.com/heroku/buildpacks-go/actions/runs/9796334219/job/27051359595#step:8:10

These are due to the fact that the `pub-hk-ubuntu-22.04-arm-small`
ARM runner also has a static IP enabled, which means all traffic from
any runners in that group go through a single load balance with the
fixed IP. This not only causes throughput/reliability issues, but means
that we're more likely to hit the Docker Hub rate limit.

As such, we requested the creation of some new runner groups that
separate out the "need a fixed IP for writes" vs "need more compute
or an ARM CPU" use-cases. As an added bonus these groups are also
using Ubuntu 24.04. See:
https://salesforce.quip.com/bu6UA0KImOxJ#temp:C:GZR57becb9df8d94f80b132168fd

This PR switches to those new groups, and also adds pre-pulling of
the run image to prevent it being pulled multiple times when the
tests run (`libcnb-test` uses `--pull-policy if-not-present` with
`pack build`, but the tests run concurrently so on first run multiple
pulls will otherwise be performed). See also:
https://github.com/heroku/buildpacks-python/pull/223

GUS-W-16238120.