Enable User env vars in subprocesses
When `clear-env` is true or not present, any environment variables in `$CNB_PLATFORM_DIR/env` will be copied into the process' current environment variables. When it is set to false, the values are still on disk but not copied over.

To support letting a customer pass information to a subcommand such as `go install,` we can set `clear-env` to its default state by removing it and passing on all current environment variables to subprocesses.